+++ bindings/nostr-ffi/src/types/filter.rs
@@ -406,8 +406,7 @@ impl Filter {
     }
 
     /// Determine if `Filter` match given `Event`.
-    ///
-    /// The `search` filed is not supported yet!
+    #[inline]
     pub fn match_event(&self, event: &Event) -> bool {
         self.inner.match_event(event.deref())
     }